+++ server/.eslintrc.js
@@ -1,15 +1,15 @@
 module.exports = {
-    env: {
-        browser: true,
-        commonjs: true,
-        es2021: true,
-        node: true,
-    },
-    extends: ['google'],
-    parser: '@typescript-eslint/parser',
-    parserOptions: {
-        ecmaVersion: 13,
-    },
-    plugins: ['@typescript-eslint'],
-    rules: {},
+  env: {
+    browser: true,
+    commonjs: true,
+    es2021: true,
+    node: true,
+  },
+  extends: ['google'],
+  parser: '@typescript-eslint/parser',
+  parserOptions: {
+    ecmaVersion: 13,
+  },
+  plugins: ['@typescript-eslint'],
+  rules: {},
 };
