+++ server/index.ts
@@ -6,8 +6,8 @@ import logger from './util/logger';
  */
 const port = app.get('PORT');
 const server = app.listen(port, () => {
-    logger.info(`Server listening on port ${port}`);
-    logger.info('ctrl + C to kill it');
+  logger.info(`Server listening on port ${port}`);
+  logger.info('ctrl + C to kill it');
 });
 
 export default server;
