+++ server/package.json
@@ -13,7 +13,7 @@
         "lint": "npx tsc --noEmit && eslint \"**/*.{js,ts}\" --quiet --fix",
         "copy-static": "cp -R public dist/ && cp package* dist/",
         "build-local": "npm install --production=false && npm run build && npm run copy-static && npm run post-build",
-        "build-dist": "npm i typescript && npm run compile && npm run copy-static && npm run post-build",
+        "build-dist": "npm i  && npm run compile && npm run copy-static && npm run post-build",
         "w-typescript": "npx tsc -w",
         "w-node": "nodemon dist/index.js",
         "dev": "concurrently -k -p \"[{name}]\" -n \"TypeScript,Node\" -c \"yellow.bold,green.bold\" \"npm run w-typescript\" \"npm run w-node\"",
@@ -28,9 +28,12 @@
     "license": "Apache-2.0",
     "dependencies": {
         "@google-cloud/logging-winston": "^4.1.1",
+        "cors": "^2.8.5",
         "dotenv": "^10.0.0",
         "express": "^4.17.1",
+        "express-session": "^1.17.2",
         "googleapis": "^91.0.0",
+        "passport": "^0.5.2",
         "passport-google-oauth20": "^2.0.0",
         "passport-google-oidc": "^0.1.0",
         "typescript": "^4.5.2",
@@ -48,7 +51,7 @@
         "@typescript-eslint/eslint-plugin": "^5.4.0",
         "@typescript-eslint/parser": "^5.4.0",
         "concurrently": "^6.4.0",
-        "eslint": "^8.3.0",
+        "eslint": "^8.6.0",
         "eslint-config-google": "^0.14.0",
         "jest": "^27.3.1",
         "nodemon": "^2.0.15",
