+++ server/routes/index.ts
@@ -14,23 +14,23 @@ router.get('/auth/login', showLogin);
 router.get(
     '/auth/google',
     passport.authenticate('google', {
-        scope: ['email', 'profile'],
-        failureRedirect: '/auth/login',
-    })
+      scope: ['email', 'profile'],
+      failureRedirect: '/auth/login',
+    }),
 );
 router.get(
     '/auth/oauthcallback',
     passport.authenticate('google', {failureRedirect: '/auth/login'}),
     (req: Request, res: Response) => {
-        res.redirect('/auth/done');
-    }
+      res.redirect('/auth/done');
+    },
 );
 // protected route
 router.get('/api/account', pass.isAuthenticated, someController.hello);
 
 router.get('/', (req: Request, res: Response) => {
-    const name = process.env.NAME || 'World';
-    res.send(`Hello ${name}! IFTTA`);
+  const name = process.env.NAME || 'World';
+  res.send(`Hello ${name}! IFTTA`);
 });
 
 export default router;
